GL-44 bugs fixed, code cleanup
* GL-44 bugs fixed, code cleanup

Closes #44

+++ development/project_urban/Assets/Scripts/EconomySystem/FoodConsumer.cs
@@ -9,7 +9,7 @@ namespace EconomySystem
     public class FoodConsumer : MonoBehaviour
     {
         public static event Action<float> ConsumeFood;
-        [SerializeField] private int foodConsumptionPerSecond = 10;
+        [SerializeField] private float foodConsumptionPerSecond = 10;
 
         public void Start()
         {
